@@ -1,11 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from '.';
 
 describe('App', () => {
     it('renders without crashing', () => {
-        const div = document.createElement('div');
-        ReactDOM.render(<App />, div);
-        ReactDOM.unmountComponentAtNode(div);
+        const container = document.createElement('div');
+        const root = createRoot(container);
+        root.render(<App />);
+        root.unmount();
     });
 });
